add test

+++ src/Resources/views/stackoverflow_details.html.php
@@ -4,17 +4,38 @@ if (!empty($message)) {
     $cache = [];
     if (file_exists($cacheFile)) {
         $cache = json_decode(file_get_contents($cacheFile), true);
+        // clean cache when it more than 20
+        if (count($cache) > 20) {
+            $cache = [];
+        }
     }
+    // if the cache exists, use it
     if (isset($cache[$message])) {
         $questions = $cache[$message];
     } else {
         try {
+            $urlMessage = urlencode($message);
+            $url = 'https://api.stackexchange.com/2.2/search/advanced'
+                . '?page=1&pagesize=5&tagged=php&order=desc&sort=relevance&q='
+                . $urlMessage
+                . '&accepted=True&site=stackoverflow';
+
             $client = new GuzzleHttp\Client();
-            $json = $client->get(
-                "https://api.stackexchange.com/2.2/search/advanced"
-                . "?page=1&pagesize=5&tagged=php&order=desc&sort=relevance&q=$message&accepted=True&site=stackoverflow"
-            )->getBody()->getContents();
-            $questions = json_decode($json, true)['items'];
+            $response = $client->get($url);
+            // Make sure API return status code with 200.
+            if ($response->getStatusCode() !== 200) {
+                throw new Exception(
+                    "Fail to access StackExchange API! The response status code return {$response->getStatusCode()}."
+                );
+            }
+
+            $apiContent = json_decode($response->getBody()->getContents(), true);
+            // Make sure get the right content of API
+            if (!is_array($apiContent) || !array_key_exists('items', $apiContent)) {
+                throw new Exception("Wrong content return from StackExchange API!");
+            }
+            $questions = $apiContent['items'];
+
             // save cache
             $cache[$message] = $questions;
             file_put_contents($cacheFile, json_encode($cache));
@@ -23,7 +44,7 @@ if (!empty($message)) {
                 $wrong = "0 results found containing '$message'";
             }
         } catch (Exception $e) {
-            $wrong = "something wrong about this package &lt;zane/whoops-stackoverflow&gt;: {$e->getMessage()}";
+            $wrong = "Something wrong when getting questions from StackOverflow: <br/><br/> {$e->getMessage()}";
         }
     }
     ?>
